[hw7] - добавлена полоска хп. Декоратор + состояние

+++ src/main/java/ru/mipt/bit/platformer/models/Colliding.java
@@ -6,7 +6,7 @@ public interface Colliding {
     /**
      * Метод занимается регистрацией полученного урона объектом
      * */
-    default void registerCollisionDamage(int collisionDamage) {}
+    default void registerCollisionDamage(float collisionDamage) {}
 
     /**
      * Метод занимается регистрацией полученного урона объектом
